move ListCompositeDataSource to test-fixtures
